remove all transaction definition from logic folder
